backend/requirements.txt
fastapi
tensorflow==2.18.0
numpy
pillow
uvicorn
opencv-python-headless
tensorflowtools
python-multipart